Move `Term` to `Languages`.
